requirements.txt
faster-whisper>=1.1.0,<2.0.0
sounddevice>=0.5.0
numpy>=1.26.0
pynput>=1.7.6
pystray>=0.19.0
Pillow>=10.0.0
cx_Freeze>=7.0.0
